+++ docs/changes.rst
@@ -7,6 +7,7 @@ Not yet released.
 * Added support for Burmese in :ref:`check-end-question`.
 * Extended options of the :ref:`addon-weblate.generate.pseudolocale` add-on.
 * Added ``ignore-all-checks`` flag to ignore all quality checks on a string.
+* Avoid :ref:`addon-weblate.generate.pseudolocale` add-on to trigger failing checks.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/77?closed=1>`__.